Bug#36178174 Incorrect query result when limit is higher than 4294967294
Changed the internal types inside MRS implementation for variables
representing limit, offset from u32 to u64.
Added tests.

MRS schema still needs update to change the type of items_on_page field.

+++ router/src/mysql_rest_service/include/mrs/database/json_template.h
@@ -43,7 +43,7 @@ class JsonTemplate {
   virtual void begin_resultset(const std::string &url,
                                const std::string &items_name,
                                const std::vector<helper::Column> &columns) = 0;
-  virtual void begin_resultset(uint32_t offset, uint32_t limit,
+  virtual void begin_resultset(uint64_t offset, uint64_t limit,
                                bool is_default_limit, const std::string &url,
                                const std::vector<helper::Column> &columns) = 0;
   virtual bool push_json_document(const char *document) = 0;
