+++ router/src/mysql_rest_service/src/mrs/json/response_sp_json_template_nest.cc
@@ -51,7 +51,7 @@ void ResponseSpJsonTemplateNest::begin_resultset(
 }
 
 void ResponseSpJsonTemplateNest::begin_resultset(
-    uint32_t, uint32_t, bool, const std::string &,
+    uint64_t, uint64_t, bool, const std::string &,
     const std::vector<helper::Column> &) {
   assert(false && "not implemented in sp");
 }
